added sign up module with frontend validation
